Fixed build

+++ bin/setup_alacritty.sh
@@ -15,8 +15,11 @@ if ! command_exists alacritty; then
 
     sudo cp target/release/alacritty /usr/local/bin
     sudo cp extra/logo/alacritty-term.svg /usr/share/pixmaps/Alacritty.svg
-    sudo desktop-file-install extra/linux/Alacritty.desktop
-    sudo update-desktop-database
+
+    if command_exists desktop-file-install; then
+        sudo desktop-file-install extra/linux/Alacritty.desktop
+        sudo update-desktop-database
+    fi
 
     success "alacritty has been installed"
 else
